Add instrument list receiver
This will send a broadcast with two arrays.  One of the arrays consists of long which correspond to the remote id of the available instruments.  The other array is list of titles of the available instruments.  The two are mapped by identical indices.  Android does not allow maps to pass as intents, which is why we must rely on the index mapping instead of just passing as a map.